python_index/requirements.txt
art==6.3
certifi==2024.8.30
charset-normalizer==3.3.2
Events==0.5
idna==3.10
opensearch-py==2.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.3
